# Block offsets that wrapped past two gigabytes, and writers that wrote what fit
EFS, GFS1, HTFS and JFS1 multiplied a block index by the block size in ints.
A couple of gigabytes in, the product wrapped: the writer placed a payload at
a negative offset and the reader fetched a file from wherever the wrap landed,
so it came back the right length holding another file's bytes. All four now
compute the offset in 64 bits.

The in-place modifiers read the whole volume into an array to walk their
trees, which no volume past two gigabytes fits in. Twenty-six filesystems now
route an add or remove above that size through an unpack-and-relay rebuild
that uses the format's own streaming writer.

FAT's rebuild-based add laid a file out as far as the cluster heap reached and
dropped the rest, so it read back short. The writer now refuses a payload
larger than the volume, on both its buffered and streaming paths, and the
rebuild stages the new volume in scratch — truncating the archive first meant
a refused add cost the caller everything already on the disk.

exFAT's defragmentation fallback finished a streaming layout with the
buffered finaliser, which writes only the entries added as arrays: the volume
came back the right shape with every file's contents missing.

The defragmentation planner resolved move dependencies in time quadratic per
pass with a quadratic pass limit. A volume fragmented into thousands of runs
turned that into a planning run that never finished; the pass limit is now
linear and a plan larger than the planner resolves is refused so the caller
rebuilds instead.

D71, D81 and T64 ended a file quietly when the medium filled — the directory
still claimed it, and a truncated D71 chain read back as a loop. Together with
DragonFS, Human68k, Apple Pascal, PC-98, LIF, Minix, Coherent, MFS and TR-DOS,
each now names what the medium holds and what was asked of it.

Tux2, Tux3, Jffs2 and FATX could not re-lay their volumes out: the first three
fed their own synthetic entries — the raw image and the metadata sheets —
back in as files, and FATX built the image in an array.

+++ FileSystems/FileSystem.Efs/EfsReader.cs
@@ -44,7 +44,10 @@ public sealed class EfsReader {
     if (entry.IsDirectory) return [];
     var len = entry.Size;
     if (len == 0 || entry.FirstBlock == 0) return [];
-    var start = entry.FirstBlock * EfsWriter.BasicBlock;
+    // Block index times block size in ints wraps a couple of gigabytes in, and
+    // the file was then read from the wrong offset — the bytes came back the
+    // right length holding something else.
+    var start = (long)entry.FirstBlock * EfsWriter.BasicBlock;
     if (start + len > _image.Length)
       throw new InvalidDataException("EFS extract: extent reaches past image end.");
     return _image.Read(start, len);
@@ -54,7 +57,7 @@ public sealed class EfsReader {
     var info = ReadInode(inode);
     if (!info.IsDirectory) return;
     if (info.NumExtents == 0 || info.FirstBlock == 0) return;
-    var off = info.FirstBlock * EfsWriter.BasicBlock;
+    var off = (long)info.FirstBlock * EfsWriter.BasicBlock;
     if (off >= _image.Length) return;
     var blk = _image.Read(off, (int)Math.Min(EfsWriter.BasicBlock, _image.Length - off));
     if (blk.Length < 3) return;
